refactor(blocks): fold BlockContentValidationError into validation-errors.ts
Senpai re-review flagged the cmd-P collision between the new
`validation-error.ts` (singular) and existing `validation-errors.ts`
(plural) one directory apart. Folding the class into the plural file
keeps the typed-issue shape and its named-error wrapper colocated, and
the file already imports `BlockContentValidationResult` so the new
`fromResult` factory sits next to its result-type input.
